requirements.txt
flask
torch
torchvision
Pillow
gunicorn
python-dotenv
timm
numpy